requirements.txt
python-telegram-bot==11.1.0
dialogflow==0.6.0
